backend/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
black==23.12.1
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
Django==4.2.9
django-filter==23.5
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-extra-fields==3.7.0
filetype==1.2.0
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
more-itertools==10.2.0
msgpack==1.0.7
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
phonenumberslite==8.13.27
Pillow==9.5.0
pip==23.3.2
platformdirs==4.1.0
pluggy==0.13.1
py==1.11.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
pytest==5.4.3
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
service-identity===24.1.0
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
urllib3==2.1.0
wcwidth==0.2.13
zope.interface==6.1